+++ app/src/test/java/com/weighttrack/ui/ChartAccessibilityTest.kt
@@ -63,6 +63,7 @@ class ChartAccessibilityTest {
                         marker = Color.White,
                         markerSurface = Color.Black,
                         markerText = Color.White,
+                        waterBand = Color.Magenta,
                     ),
                 )
             }